Fix a broken assumption in the Socket tests.
SocketClient assumed that it was safe to dispose the server socket if
its call to Connect() failed. This is not necessarily the case, however:
the calls to Accept and Connect race, there is a chance that the dispose
happens before the call to Accept. In this case, the call to Accept will
throw ObjectDisposedException and fail the test. Work around this issue
by changing SocketClient to wait 100 milliseconds before disposing the
server socket.